requirements.txt
# Core dependencies
docling==2.43.0
langchain-text-splitters>=0.3.0
google-generativeai>=0.8.0
google-genai>=0.11.0
httpx>=0.25.0
faiss-cpu>=1.7.4
transformers>=4.40.0
accelerate>=0.30.0
torch>=2.0.0
cohere>=5.0.0
python-dotenv>=1.0.0
numpy>=1.21.0

# Document processing dependencies
PyPDF2>=3.0.0
python-docx>=1.1.0

# FastAPI dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
